+++ Game/Scripts/UIComboManager.cpp
@@ -61,7 +61,7 @@ void UIComboManager::SetActivateSpecial(bool activate)
 	}
 }
 
-void UIComboManager::SetComboBarValue(int value)
+void UIComboManager::SetComboBarValue(float value)
 {
 	comboBar->ModifyCurrentValue(static_cast<float>(value));
 	if (value == 0)
